refactor(agent-env): slim AgentEnvPanel onto controller+store+view-model
